+++ src/main/java/com/rmg/rantMyGames/repository/IGameRepository.java
@@ -3,6 +3,6 @@ package com.rmg.rantMyGames.repository;
 import com.rmg.rantMyGames.entity.Game;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IGameRepository extends JpaRepository <Game, Long> {
+public interface IGameRepository extends JpaRepository <Game, Integer> {
     boolean existsByGameTitle(String gameTitle);
 }
