+++ src/main/java/com/rmg/rantMyGames/service/ICRUDService.java
@@ -4,8 +4,8 @@ import java.util.List;
 
 public interface ICRUDService <DTO, Entity>{
     void create(DTO dto) throws Exception;
-    DTO read(Long id) throws Exception;
-    void delete(Long id) throws Exception;
+    DTO read(Integer id) throws Exception;
+    void delete(Integer id) throws Exception;
     void update(DTO dto) throws Exception;
     List<DTO> getAll();
 }
